Fixing broken date

+++ Tasker/data/Task.swift
@@ -11,10 +11,10 @@ import RealmSwift
 class Task: Object, ObjectKeyIdentifiable {
     @Persisted var title: String? = nil
     @Persisted var taskDescription: String? = nil
-    @Persisted var deadline: Int? = nil
+    @Persisted var deadline: Double? = nil
     @Persisted var completed: Bool = false
     
-    convenience init(title: String?, taskDescription: String?, deadline: Int?, completed: Bool = false) {
+    convenience init(title: String?, taskDescription: String?, deadline: Double?, completed: Bool = false) {
         self.init()
         self.title = title
         self.taskDescription = taskDescription
